requirements.txt
sqlalchemy
geoalchemy2
psycopg2-binary
flask
flask-cors
fire
tabulate
progressbar2
overpy
shapely>=1.8.0
requests
beautifulsoup4
lxml
python-dateutil
pandas
stdiomask
simple_term_menu
streamlit
streamlit_ace
extra-streamlit-components
universal-analytics-python3
